Create model scripts and code for RUTU
This commit creates the model scripts and code for 3 different versions of a GLMM used previously for Red Knots, but now applied to Ruddy Turnstones (RUTU). These 3 different versions include an examination of the relationships betweeen RUTU and planes, RUTU and Tending, and the additive effect of RUTU and Tending and the number of gulls.